requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.64
botocore==1.40.64
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
deepface==0.0.90
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.119.0
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flatbuffers==25.9.23
frozenlist==1.8.0
gast==0.4.0
gdown==5.2.0
google-auth==2.42.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
keras==2.11.0
libclang==18.1.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==1.0.0
multidict==6.7.0
namex==0.1.0
numpy==1.26.0
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==3.19.6
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
srt==3.5.3
starlette==0.48.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tqdm==4.67.1
twilio==9.8.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vonage==4.7.2
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.6.2
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.7
vonage-numbers==1.0.5
vonage-sms==1.1.6
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.5.1
vonage-voice==1.4.0
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
uvicorn[standard]

# Google Maps integration
googlemaps==4.10.0
# CORS middleware
fastapi-cors==0.0.6

# HTTP requests (for testing)
httpx==0.25.1

# Text Detection / OCR
easyocr==1.7.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1